+++ test/common/util.test.ts
@@ -1,8 +1,8 @@
 import {expect, test} from '@oclif/test'
 import * as YAML from 'yaml'
-import {isYamlSame} from '../src/util'
+import {isYamlSame} from '../../src/common/util'
 import * as fs from 'fs'
-import {ArrayParser} from '../src/util'
+import {ArrayParser} from '../../src/common/util'
 
 describe('isYamlSame', () => {
   test
